Ponderada4/app/requirements.txt
pycaret[full]
fastapi
uvicorn
psycopg2-binary
matplotlib
pandas
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv